Add files via upload
The FER-2013 dataset is widely used in the field of emotion detection and facial expression recognition. It contains grayscale images of faces, each of which is labeled with one of seven emotions: angry, disgust, fear, happy, sad, surprise, and neutral. The dataset is composed of 35,887 images of size 48x48 pixels, split into training and validation sets. The images were collected using the Google image search API, and each face is centered and occupies a majority of the image area. The dataset is challenging due to variations in lighting, facial features, and expressions, making it a suitable choice for building and evaluating robust models.